refactor PE section filenames
