+++ akvo/rsr/spa/app/modules/editor/section2/section2.jsx
