+++ akvo/rsr/spa/app/modules/editor/section3/section3.jsx
