+++ akvo/rsr/spa/app/modules/editor/section7/section7.jsx
